pytest: Separate client api tests
